Stability improvements

+++ admin/enablefees.php
@@ -31,8 +31,8 @@ if (isset($_POST['action']) && $_POST['action'] == 'update')
 	// Update database
 	$system->writesetting("fees", $_POST['fees'], "str");
 	$system->writesetting("fee_type", $_POST['fee_type'], "int");
-	$system->writesetting("fee_max_debt", $system->input_money($_POST['fee_max_debt']), "int");
-	$system->writesetting("fee_signup_bonus", $system->input_money($_POST['fee_signup_bonus']), "int");
+	$system->writesetting("fee_max_debt", $system->input_money($_POST['fee_max_debt']), "float");
+	$system->writesetting("fee_signup_bonus", $system->input_money($_POST['fee_signup_bonus']), "float");
 	$system->writesetting("fee_disable_acc", $_POST['fee_disable_acc'], "str");
 	$system->writesetting("payment_gateway_sandbox", $_POST['payment_gateway_sandbox'], "str");
